+++ src/xtd.forms/src/xtd/forms/context_menu.cpp
@@ -63,7 +63,7 @@ void context_menu::on_item_removed(size_t pos, menu_item_ref item) {
   native::context_menu::remove_item(handle(), pos);
 }
 
-bool context_menu::on_item_click(int32_t menu_id) {
+bool context_menu::on_item_click(intptr_t menu_id) {
   auto it = handles_.find(menu_id);
   if (it != handles_.end()) {
     auto& menu = static_cast<menu_item&>(it->second.get());
